+++ projects/tuist/fixtures/ios_app_with_transitive_project/App/Sources/AppDelegate.swift
@@ -1,4 +1,5 @@
 import Framework1
+import FrameworkA
 import UIKit
 
 @UIApplicationMain
@@ -6,11 +7,16 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     var window: UIWindow?
 
     func applicationDidFinishLaunching(_: UIApplication) {
-        let framework1 = Framework1File()
-
         print(hello())
+
+        let framework1 = Framework1File()
         print("AppDelegate -> \(framework1.hello())")
         print("AppDelegate -> \(framework1.helloFromFramework2())")
+
+        let frameworkA = FrameworkAFile()
+        print("AppDelegate -> \(frameworkA.hello())")
+        print("AppDelegate -> \(frameworkA.helloFromFrameworkB())")
+        print("AppDelegate -> \(frameworkA.helloFromFrameworkC())")
     }
 
     func hello() -> String {