+++ projects/tuist/fixtures/ios_app_with_transitive_project/FrameworkA/Sources/FrameworkAFile.swift
@@ -1,0 +1,20 @@
+import Foundation
+import FrameworkB
+
+public class FrameworkAFile {
+    private let frameworkBFile = FrameworkBFile()
+
+    public init() {}
+
+    public func hello() -> String {
+        "FrameworkAFile.hello()"
+    }
+
+    public func helloFromFrameworkB() -> String {
+        "FrameworkAFile -> \(frameworkBFile.hello())"
+    }
+
+    public func helloFromFrameworkC() -> String {
+        "FrameworkAFile -> \(frameworkBFile.helloFromFrameworkC())"
+    }
+}