+++ projects/tuist/fixtures/ios_app_with_transitive_project/FrameworkB/Sources/FrameworkBFile.swift
@@ -1,0 +1,16 @@
+import Foundation
+import FrameworkC
+
+public class FrameworkBFile {
+    private let frameworkCFile = FrameworkCFile()
+
+    public init() {}
+
+    public func hello() -> String {
+        "FrameworkBFile.hello()"
+    }
+
+    public func helloFromFrameworkC() -> String {
+        "FrameworkBFile -> \(frameworkCFile.hello())"
+    }
+}